requirements.txt
absl-py==2.3.1

aiofiles==24.1.0
annotated-types==0.7.0
asttokens==3.0.0
astunparse==1.6.3
blinker==1.9.0
Brotli==1.1.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm==0.2.3
debugpy==1.8.15
decorator==5.2.1
exceptiongroup==1.3.0
executing==2.2.0
ffmpy==0.6.1
filelock==3.18.0
Flask==3.1.1
flatbuffers==25.2.10
fsspec==2025.7.0
gast==0.6.0
google-pasta==0.2.0
groovy==0.1.2
grpcio==1.74.0
h11==0.16.0
h5py==3.14.0
idna==3.10
ipykernel==6.30.0
ipython==8.25.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.8.1

libclang==18.1.1

markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
namex==0.1.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.1.3
opt_einsum==3.4.0
optree==0.17.0
orjson==3.11.1
packaging==25.0
pandas==2.3.1
parso==0.8.4
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
protobuf==5.29.5
psutil==7.0.0
pure_eval==0.2.3
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
pywin32==311; sys_platform == "win32"
PyYAML==6.0.2
pyzmq==27.0.0
requests==2.32.4
rich==14.1.0
ruff==0.12.5
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2


termcolor==3.1.0
tomlkit==0.13.3
torch==2.7.1
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
vision==1.0.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
